requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
bson==0.5.8
dataclasses==0.7
Django==2.2.9
django-jsonfield==1.4.0
djangorestframework==3.11.0
djongo==1.2.35
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
pymongo==3.10.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.7
wrapt==1.11.2
